+++ pico_limbo/src/configuration/experimental.rs
@@ -1,0 +1,19 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Default)]
+pub struct ExperimentalConfig {
+    pub world: ExperimentalWorldConfig,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct ExperimentalWorldConfig {
+    pub spawn_position: (f64, f64, f64),
+}
+
+impl Default for ExperimentalWorldConfig {
+    fn default() -> Self {
+        Self {
+            spawn_position: (0.0, 320.0, 0.0),
+        }
+    }
+}