+++ pico_limbo/src/configuration/forwarding.rs
@@ -1,19 +1,19 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Default)]
+#[derive(Serialize, Deserialize, Default)]
 pub struct ModernForwardingConfig {
     pub enabled: bool,
     pub secret: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Default)]
+#[derive(Serialize, Deserialize, Default)]
 pub struct BungeeCordForwardingConfig {
     pub enabled: bool,
     pub bungee_guard: bool,
     pub tokens: Vec<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Default)]
+#[derive(Serialize, Deserialize, Default)]
 pub struct ForwardingConfig {
     pub velocity: ModernForwardingConfig,
     pub bungee_cord: BungeeCordForwardingConfig,