+++ pico_limbo/src/configuration/game_mode_config.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Deserialize, Serialize, Default)]
+#[derive(Deserialize, Serialize, Default)]
 #[serde(rename_all = "lowercase")]
 pub enum GameModeConfig {
     Survival,